Sort best matching movie results by popularity
When searching movies based solely on title and year, there may be
multiple hits as seen in commit: 80c27ac

This commit make sure to order those multiple best matching results by
popularity. This should make sure user gets the most likely result
first.

Results for 'Dragons' + '2010' + fr-FR will be in that order:

1. https://www.themoviedb.org/movie/10191-how-to-train-your-dragon?language=fr-FR
2. https://www.themoviedb.org/movie/40911?language=fr-FR

Documentation at:

* https://developers.themoviedb.org/3/search/search-movies

indicates that 'popularity' (number) is an optional response key, the
code defaults to 0 when key is not present.